backend/requirements.txt
fastapi
uvicorn[standard]
websockets
openai
pydantic
python-dotenv
httpx
readability-lxml
